==> Scripts/14-Beta_diversity.Rmd <==
---
title: "Beta_diversity"
author: "Patrick Pata"
date: "26/10/2021"
output: html_document
---

This creates figure 6 and supplementary figure S6 of the Pata et al. PinO manuscript.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load files and libraries
```{r}
ls.packages <- c("tidyverse", "vegan","adespatial","ggrepel","cowplot",
                 "rnaturalearth","tictoc","reshape2","dunn.test","here",
                 "ggpubr")
new.packages <- ls.packages[!(ls.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)

library(tidyverse)
library(vegan)
library(adespatial)
library(ggrepel)
library(cowplot)
library(rnaturalearth)
library(tictoc)
library(reshape2)
library(dunn.test)
library(here)
library(ggpubr)
`%notin%` <- Negate(`%in%`)

load(here::here("Data_Output/zoop_data_regionalized_10252021.RData"))
theme_set(theme_bw())
```

# 1. Species contribution and local contribution to beta-diversity.
This is calculated using the transformed community composition matrix.
```{r}
betadiv.CBD <- beta.div(zoop_abundance, method = "log.chord", nperm = 9999, clock = TRUE)
summary(betadiv.CBD)
betadiv.CBD$beta # SSTotal and BDTotal

# Which species have a SCBD larger than the mean SCBD?
betadiv.CBD$SCBD[betadiv.CBD$SCBD > mean(betadiv.CBD$SCBD)]

# how much of the total BD do the species represent
sum(betadiv.CBD$SCBD[betadiv.CBD$SCBD > mean(betadiv.CBD$SCBD)]) / sum(betadiv.CBD$SCBD)
```

Correlate SCBD with species presence and mean abundance
SCBD - species contribution to total beta diversity.
```{r, fig.width=8, fig.height=10.5}
bd.SCBD <- as.data.frame(betadiv.CBD$SCBD) %>% 
  rownames_to_column("Species") %>% 
  as_tibble() %>% 
  rename(SCBD = "betadiv.CBD$SCBD") %>% 
  left_join(sp.abund.list)

cor.test(bd.SCBD$SCBD, bd.SCBD$presence.perc)
cor.test(bd.SCBD$SCBD, bd.SCBD$abundance.mean)

# Update sp names of some species
bd.SCBD$Species[bd.SCBD$Species == "Nematoscelis"] <- "Nematoscelis difficilis"
bd.SCBD$Species[bd.SCBD$Species == "Alacia pseudoalata"] <- "Alacia major"
bd.SCBD$Species[bd.SCBD$Species == "Metridia pseudopacifica"] <- "Metridia aff. lucens"
bd.SCBD$Species[bd.SCBD$Species == "Orthoconchoecia striola"] <- "Orthoconchoecia aff. striola"
bd.SCBD$Species[bd.SCBD$Species == "Mesosagitta decipiens"] <- "Decipisagitta decipiens"
bd.SCBD$Species[bd.SCBD$Species == "Aglantha"] <- "Aglantha digitale"

# Identify species names to plot and font style
splabels <- dplyr::filter(bd.SCBD, SCBD > mean(SCBD)) %>% 
  left_join(dplyr::select(taxalist, Species = Taxa.Name, Taxa.Level)) %>% 
  mutate(Taxa.Level = if_else(Taxa.Level %in% c("C","O","F"),"plain","italic")) %>% 
  unique()


g1 <- ggplot(bd.SCBD, aes(x = SCBD, y = presence.perc)) + 
  geom_point(alpha = 0.6) +
  geom_text_repel(data = splabels,
                  aes(x = SCBD, y = presence.perc, label = Species, 
                      fontface = Taxa.Level), max.overlaps = 15) +
  theme_bw() +
  ylab("% Presence in samples") +
  theme(legend.text = element_text(size = 16),
        axis.title = element_text(size = 16),
        axis.text = element_text(size = 16))


g2 <- ggplot(bd.SCBD, aes(x = SCBD, y = abundance.mean)) + 
  geom_point() +
  theme_bw() +
  scale_y_continuous(trans='log10') +
  ylab("Abundance") +
  theme(legend.text = element_text(size = 16),
        axis.title = element_text(size = 16),
        axis.text = element_text(size = 16))

plot_grid(g1,g2, nrow = 2, align = "v", rel_heights = c(3.5,1), labels = c("A","B"))
```

Figure 6. SCBD but without species names, change symbol for species > mean SCBD
```{r, fig.height=7, fig.width=5}
g1 <- ggplot() + 
  geom_point(data = dplyr::filter(bd.SCBD, SCBD <= mean(SCBD)), 
             aes(x = SCBD, y = presence.perc), shape = 1) +
  geom_point(data = dplyr::filter(bd.SCBD, SCBD > mean(SCBD)), 
             aes(x = SCBD, y = presence.perc), shape = 16) +
  # geom_vline(xintercept = mean(bd.SCBD$SCBD), linetype = "dashed")+
  theme_bw() +
  scale_x_continuous(expand = c(0.0001, 0.001)) +
  ylab("% Presence in samples") +
  theme(legend.text = element_text(size = 12),
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 12),
        axis.text.x=element_blank(),
        axis.title.x=element_blank())

g2 <- ggplot() + 
  geom_point(data = dplyr::filter(bd.SCBD, SCBD <= mean(SCBD)), 
             aes(x = SCBD, y = abundance.mean), shape = 1) +
  geom_point(data = dplyr::filter(bd.SCBD, SCBD > mean(SCBD)), 
             aes(x = SCBD, y = abundance.mean), shape = 16) +
  # geom_vline(xintercept = mean(bd.SCBD$SCBD), linetype = "dashed")+
  theme_bw() +
  scale_x_continuous(expand = c(0.0001, 0.001)) +
  scale_y_continuous(trans='log10') +
  ylab("Abundance (indivs m\u207B\u00B2)") +
  xlab("Species contribution to beta-diversity") +
  theme(legend.text = element_text(size = 12),
        axis.title = element_text(size = 12),
        axis.text = element_text(size = 12))

png(paste0("Figures/Figure_6_",Sys.Date(),".png"),
    width = 5, height = 6, units = "in", res = 300)
plot_grid(g1,g2, nrow = 2, align = "v", rel_heights = c(1,1.1), 
          labels = c("A","B"))
dev.off()
```


Local contribution to species diversity (LCBD)
$p.adj is the Holm corrected p-value for LCBD indices. 
```{r, eval = FALSE}
# After holm correction, none of the sites have significant LCBDs. So the p.LCBD is used here as a filter instead to visualize the points.
bd.LCBD <- as.data.frame(betadiv.CBD[c(3:5)]) %>% 
  rownames_to_column("Key") %>% 
  as_tibble() %>% 
  left_join(dplyr::select(zoopmeta,Key,Bioregion,Longitude,Latitude,DateTime,Year)) %>% 
  left_join(dplyr::select(zoopenvi,Key,DayOfYear)) %>% 
  filter(p.LCBD < 0.05)

# Spatial
world <- ne_countries(scale = "medium", returnclass = "sf")
ggplot(data = world) +
    geom_sf() +
    geom_point(data = bd.LCBD, aes(x = Longitude, y = Latitude, 
                                   size = LCBD), shape = 1) +
    coord_sf(xlim = c(-145.1, -122), ylim = c(47, 56), expand = FALSE) +
    xlab("Longitude") + ylab("Latitude") +
    theme_bw()

# Temporal
ggplot(data = bd.LCBD, aes(x = DayOfYear, y = Year, size = LCBD )) +
  geom_point( shape = 1) +
  theme_bw()
```


# 2. SIMPER
Species contribution to beta diversity according to groupings. Note that this is on BCD. Pairs of groups are compared. 
```{r}
tic(simper)
bd.simper <- simper(zoop.log, bioregions$Bioregion )
toc()

summary(bd.simper)
```

Extracting SIMPER results
```{r}
# Get overall dissimilarity between clusters
bd.simper.overall <- as_tibble(names(bd.simper)) %>% 
  rename(pair = "value") %>% 
  mutate(dissim = NA)
for (k in c(1:6)) {
  bd.simper.overall$dissim[k] <-  bd.simper[[k]]$overall
}
bd.simper.overall

# Get pairwise dissimilarity for each species
bd.simper.pairs <- tibble()
for (k in c(1:6)) {
  A <- cbind(bd.simper[[k]]$species, bd.simper[[k]]$average, bd.simper[[k]]$sd,
             bd.simper[[k]]$ratio, bd.simper[[k]]$ava, bd.simper[[k]]$avb,
             bd.simper[[k]]$ord, bd.simper[[k]]$cusum)
  A <- as_tibble(A)
  colnames(A) <- names(bd.simper[[k]][c(1,2,4:9)])
  A <- A %>% 
    add_column(pair = names(bd.simper[k]), index = c(1:160), .before = "species")
  if (k == 1){
    bd.simper.pairs <- A
  } else {
    bd.simper.pairs <- bind_rows(bd.simper.pairs, A)
  }
}
bd.simper.pairs[,4:10] <- sapply(bd.simper.pairs[,4:10], as.numeric)
```

Correlation between SCBD and SIMPER pairs
```{r}
bd.simper.pairs <- bd.simper.pairs %>% 
  left_join(dplyr::select(bd.SCBD, species = Species, SCBD))
cor.test(bd.simper.pairs$average, bd.simper.pairs$SCBD)

# correlate for each pair
bd.simper.pairs %>% 
  group_by(pair) %>% 
  summarise(r = cor.test(average, SCBD)$estimate, 
            pval = cor.test(average, SCBD)$p.value)
```

Supplementary Figure S6. SCBD-SIMPER correlation
```{r}
bd.simper.pairs <- bd.simper.pairs %>% 
  mutate(pair = str_replace(pair, pattern = "_", replacement = " & "))
g1 <- ggpubr::ggscatter(bd.simper.pairs,
                    x = "average", y = "SCBD",color = "pair", add = "reg.line") +
  theme_bw()+
  xlab("Average dissimilarity") +
  labs("Bioregion pair") +
  theme(legend.position="none", 
        legend.background = element_rect(colour="gray", size=0.5, linetype="solid"),
        legend.title = element_text(color = 'black', size = 16, face = 'bold'),
        legend.text = element_text(size = 16),
        axis.title = element_text(size = 16),
        axis.text = element_text(size = 16)) +
  ggpubr::stat_cor(method = "pearson", 
                   r.accuracy = 0.01, p.accuracy = 0.001) +
  facet_wrap(facets = "pair", nrow = 3) 

g1

png(paste0("Figures/Supp_Figure_S6_",Sys.Date(),".png"),
    width = 8, height = 6, units = "in", res = 300)
g1
dev.off()
```
